+++ src/V1/ServiceBricks.Storage.EntityFrameworkCore/Model/EntityFrameworkCoreStorageRepository.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
 using ServiceQuery;
 
 namespace ServiceBricks.Storage.EntityFrameworkCore
@@ -20,9 +19,12 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
         /// </summary>
         /// <param name="logFactory"></param>
         /// <param name="businessRuleService"></param>
-        public EntityFrameworkCoreStorageRepository(ILoggerFactory logFactory)
+        public EntityFrameworkCoreStorageRepository(
+            ILoggerFactory logFactory)
         {
             _logger = logFactory.CreateLogger<EntityFrameworkCoreStorageRepository<TDomain>>();
+            LogExceptions = true;
+            LogServiceQueryErrors = false;
         }
 
         /// <summary>
@@ -44,6 +46,11 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
         /// Determines if service query errors are logged.
         /// </summary>
         public virtual bool LogServiceQueryErrors { get; set; }
+
+        /// <summary>
+        /// Determines if logging is enabled.
+        /// </summary>
+        public virtual bool LogExceptions { get; set; }
 
         /// <summary>
         /// Get the storage repository.
@@ -82,7 +89,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(SaveChanges)} {ex.Message}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(SaveChanges)} {ex.Message}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -101,7 +109,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(SaveChangesAsync)} {ex.Message}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(SaveChangesAsync)} {ex.Message}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -128,7 +137,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(Delete)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(Delete)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -155,7 +165,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(DeleteAsync)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(DeleteAsync)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -177,7 +188,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(Create)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(Create)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -199,7 +211,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(CreateAsync)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(CreateAsync)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -225,7 +238,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(Update)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(Update)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -251,7 +265,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(UpdateAsync)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(UpdateAsync)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 resp.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return resp;
@@ -275,7 +290,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(GetAsync)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(GetAsync)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 response.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return response;
@@ -299,7 +315,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(GetAsync)} {ex.Message} {JsonConvert.SerializeObject(obj)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(GetAsync)} {ex.Message} {JsonSerializer.Instance.SerializeObject(obj)}");
                 response.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return response;
@@ -317,7 +334,8 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, nameof(GetQueryable));
+                if (LogExceptions)
+                    _logger.LogError(ex, nameof(GetQueryable));
             }
             return null;
         }
@@ -344,12 +362,13 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             catch (ServiceQueryException sqe)
             {
                 if (LogServiceQueryErrors)
-                    _logger.LogError(sqe, $"{nameof(Query)} {sqe.Message} {JsonConvert.SerializeObject(request)}");
+                    _logger.LogError(sqe, $"{nameof(Query)} {sqe.Message} {JsonSerializer.Instance.SerializeObject(request)}");
                 response.AddMessage(ResponseMessage.CreateError(sqe, LocalizationResource.ERROR_SERVICEQUERY));
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(Query)} {ex.Message} {JsonConvert.SerializeObject(request)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(Query)} {ex.Message} {JsonSerializer.Instance.SerializeObject(request)}");
                 response.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return response;
@@ -377,12 +396,13 @@ namespace ServiceBricks.Storage.EntityFrameworkCore
             catch (ServiceQueryException sqe)
             {
                 if (LogServiceQueryErrors)
-                    _logger.LogError(sqe, $"{nameof(QueryAsync)} {sqe.Message} {JsonConvert.SerializeObject(request)}");
+                    _logger.LogError(sqe, $"{nameof(QueryAsync)} {sqe.Message} {JsonSerializer.Instance.SerializeObject(request)}");
                 response.AddMessage(ResponseMessage.CreateError(sqe, LocalizationResource.ERROR_SERVICEQUERY));
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, $"{nameof(QueryAsync)} {ex.Message} {JsonConvert.SerializeObject(request)}");
+                if (LogExceptions)
+                    _logger.LogError(ex, $"{nameof(QueryAsync)} {ex.Message} {JsonSerializer.Instance.SerializeObject(request)}");
                 response.AddMessage(ResponseMessage.CreateError(ex, LocalizationResource.ERROR_STORAGE));
             }
             return response;