+++ src/V1/ServiceBricks/Api/ApiController.cs
@@ -11,18 +11,17 @@ namespace ServiceBricks
     /// By default, no security policy is added to this base class. Use AdminPolicyRestApiController instead.
     /// </summary>
     /// <typeparam name="TDto"></typeparam>
-    public partial class ApiController<TDto> : ControllerBase, IApiController<TDto>
+    public partial class ApiController<TDto> : BaseController, IApiController<TDto>
         where TDto : class
     {
         protected readonly IApiService<TDto> _domainObjectService;
-        protected readonly ApiOptions _apiOptions;
 
         public ApiController(
             IApiService<TDto> domainObjectService,
             IOptions<ApiOptions> apiOptions)
+            : base(apiOptions)
         {
             _domainObjectService = domainObjectService;
-            _apiOptions = apiOptions.Value ?? new ApiOptions();
         }
 
         /// <summary>
@@ -337,42 +336,6 @@ namespace ServiceBricks
                     return Ok(resp.Item);
             }
             return GetErrorResponse(resp);
-        }
-
-        /// <summary>
-        /// Get the error response
-        /// </summary>
-        /// <param name="response"></param>
-        /// <returns></returns>
-        [NonAction]
-        public virtual ActionResult GetErrorResponse(IResponse response)
-        {
-            // Create filtered list of messages (Don't expose sensitive system errors)
-            List<IResponseMessage> messages = response.Messages.ToList();
-            if (!_apiOptions.ExposeSystemErrors)
-                messages = messages.Where(x => x.Severity != ResponseSeverity.ErrorSystemSensitive).ToList();
-            if (messages.Count == 0)
-                messages.Add(ResponseMessage.CreateError(LocalizationResource.ERROR_SYSTEM));
-
-            // Create new error response
-            var respError = new Response() { Error = true };
-            foreach (var item in messages)
-                respError.AddMessage(item);
-
-            // Return response
-            if (_apiOptions.ReturnResponseObject)
-                return BadRequest(respError);
-            else
-            {
-                string detail = string.Join(Environment.NewLine, messages.Select(x => x.Message));
-                var details = new ProblemDetails()
-                {
-                    Title = LocalizationResource.ERROR_SYSTEM,
-                    Status = (int)HttpStatusCode.BadRequest,
-                    Detail = detail
-                };
-                return BadRequest(details);
-            }
         }
     }
 }