+++ src/V1/ServiceBricks/Extension/ServiceCollectionExtensions.cs
@@ -1,9 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc.ApplicationParts;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using System.Reflection;
-using System.Linq;
 
 namespace ServiceBricks
 {