+++ src/V1/ServiceBricks/Interface/Model/IJsonSerializer.cs
@@ -1,0 +1,11 @@
+﻿namespace ServiceBricks
+{
+    public partial interface IJsonSerializer
+    {
+        string SerializeObject(object obj);
+
+        T DeserializeObject<T>(string data);
+
+        object DeserializeObject(string data, Type objectType);
+    }
+}