+++ src/V1/Tests/WebAppNet6/Mapping/ProblemDetailsProfile.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 using Microsoft.AspNetCore.Mvc;
-using Newtonsoft.Json;
 using System.Net;
+using Newtonsoft.Json;
 
 namespace WebApp.Mapping
 {